+++ CHANGELOG.md
@@ -80,6 +80,7 @@ Developers may need to change their code
   
 ### Minor features
 
+* Removed `JSON_CDATA_STRIP` compile-time option (obsolete since 5.4)
 * Added fuzz code for xpath
 * List-pagination: Follow ietf-draft 100%: Removed list-pagination "presence"
 * Main example: Removed dependency of external IETF RFCs